Began transitioning to linter-clang-plus.
I'll be using some of the techniques I've learned from writing linter-swiftc an linter-crystal to create a nice, fast, and extensible version of linter-clang.

Also, because the original package has had some issues with comming up with recent updates and the maintainer seems not to be around much anymore this will serve as a method of keeping the linter up-to-date.